Refactor models to separate files, restructure
- moved models to separate files for clarity/structure
