+++ csharp/src/Bargreen.API/Controllers/InventoryController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Bargreen.Services;
+using Bargreen.Services.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -16,7 +17,7 @@ namespace Bargreen.API.Controllers
     {
         [Route("InventoryBalances")]
         [HttpGet]
-        public IEnumerable<InventoryBalance> GetInventoryBalances()
+        public async Task<InventoryBalance> GetInventoryBalances()
         {
             var inventoryService = new InventoryService();
             return inventoryService.GetInventoryBalances();
@@ -24,7 +25,7 @@ namespace Bargreen.API.Controllers
 
         [Route("AccountingBalances")]
         [HttpGet]
-        public IEnumerable<AccountingBalance> GetAccountingBalances()
+        public async Task<IEnumerable<AccountingBalance>> GetAccountingBalances()
         {
             var inventoryService = new InventoryService();
             return inventoryService.GetAccountingBalances();
@@ -32,7 +33,7 @@ namespace Bargreen.API.Controllers
 
         [Route("InventoryReconciliation")]
         [HttpGet]
-        public IEnumerable<InventoryReconciliationResult> GetReconciliation()
+        public async Task<IEnumerable<InventoryReconciliationResult>> GetReconciliation()
         {
             var inventoryService = new InventoryService();
             return InventoryService.ReconcileInventoryToAccounting(inventoryService.GetInventoryBalances(), inventoryService.GetAccountingBalances());
