+++ csharp/src/Bargreen.Services/Bargreen.Services.csproj
@@ -4,4 +4,9 @@
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Implementations" />
+    <Folder Include="Interfaces" />
+  </ItemGroup>
+
 </Project>
