+++ csharp/src/Bargreen.Services/Models/AccountingBalance.cs
@@ -1,0 +1,9 @@
+﻿namespace Bargreen.Services.Models
+{
+    public class AccountingBalance
+    {
+        public string ItemNumber { get; set; }
+        public decimal TotalInventoryValue { get; set; }
+    }
+
+}
