+++ csharp/src/Bargreen.Services/Models/InventoryBalance.cs
@@ -1,0 +1,10 @@
+﻿namespace Bargreen.Services.Models
+{
+    public class InventoryBalance
+    {
+        public string ItemNumber { get; set; }
+        public string WarehouseLocation { get; set; }
+        public int QuantityOnHand { get; set; }
+        public decimal PricePerItem { get; set; }
+    }
+}
